+++ unicorn_binance_websocket_api/exceptions.py
@@ -37,7 +37,7 @@
 # define Python user-defined exceptions
 class StreamRecoveryError(Exception):
     """
-    Exception for non recoverable streams.
+    Exception for non-recoverable streams.
     """
     pass
 
@@ -45,5 +45,12 @@ class StreamRecoveryError(Exception):
 class UnknownExchange(Exception):
     """
     Exception if the manager class is started with an unkown exchange.
+    """
+    pass
+
+
+class Socks5ProxyConnectionError(Exception):
+    """
+    Exception if the manager class is not able to establish a connection to the socks5 proxy.
     """
     pass